backend/requirements.txt
fastapi==0.128.0
uvicorn==0.40.0
sqlmodel==0.0.31
python-jose==3.5.0
passlib==1.7.4
python-multipart==0.0.21
asyncpg==0.31.0
sqlalchemy==2.0.45
pydantic==2.12.5
typing-extensions==4.15.0